Add buttons to play and event listener to them
Earlier game was played by console, this commit adds 3 buttons with IDs
equal to player choice. In JS file I deleted game() function and added
event listener instead, so when the player click on button, choice is
saved in a variable and throw as a parameter to playRound() function
and the function executes once.